initial rework of tallies and tmc
